% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rog-logo.R
\encoding{UTF-8}
\name{rog_load_font}
\alias{rog_load_font}
\title{Load the rogtemplate font}
\usage{
rog_load_font()
}
\value{
A \link[base:character]{character} string containing the font family
name, \code{"B612 Mono"}.
}
\description{
Load the current rOpenGov font,
\href{https://fonts.google.com/specimen/B612+Mono}{B612 Mono}.
}
\examples{
rog_load_font()
}
\seealso{
\CRANpkg{sysfonts}'s \code{\link[sysfonts:font_add]{sysfonts::font_add()}} for registering fonts
and \CRANpkg{showtext}'s \code{\link[showtext:showtext_auto]{showtext::showtext_auto()}} for rendering them
in plots.

Package asset helpers:
\code{\link[=rog_badge_ropengov]{rog_badge_ropengov()}},
\code{\link[=rog_logo]{rog_logo()}}
}
\concept{assets}
