% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rog-logo.R
\encoding{UTF-8}
\name{rog_logo}
\alias{rog_logo}
\title{Create a logo for your rOpenGov package}
\usage{
rog_logo(
  pkgname,
  filename = "man/figures/logo.png",
  p_x = 1,
  p_y = 1,
  p_size = 202.6 * nchar(pkgname)^-1.008,
  overwrite = FALSE,
  favicons = TRUE
)
}
\arguments{
\item{pkgname}{Name of the package. If not supplied, the name is detected
from DESCRIPTION.}

\item{filename}{filename to save sticker}

\item{p_x}{x position for package name}

\item{p_y}{y position for package name}

\item{p_size}{font size for package name}

\item{overwrite}{Whether to overwrite the current logo. If \code{FALSE} and
\code{filename} already exists, save the new logo to a temporary file instead.}

\item{favicons}{Whether to create favicons with \CRANpkg{pkgdown}'s
\code{\link[pkgdown:build_favicons]{pkgdown::build_favicons()}} when \code{filename} is \code{"man/figures/logo.png"}.}
}
\value{
This function is called for its side effects and returns
\code{\link[base:NULL]{NULL}} invisibly.
}
\description{
Create a logo automatically with \CRANpkg{hexSticker}'s
\code{\link[hexSticker:sticker]{hexSticker::sticker()}}. Optionally, create favicons with
\CRANpkg{pkgdown}'s \code{\link[pkgdown:build_favicons]{pkgdown::build_favicons()}}.
}
\examples{
tmp <- tempfile(fileext = ".png")
rog_logo("test a package", tmp, overwrite = FALSE, favicons = FALSE)

# Display the logo.
logo <- magick::image_read(tmp)

logo

plot(logo)
}
\seealso{
\code{\link[=rog_build]{rog_build()}} to create the logo as part of a local site build.
\CRANpkg{hexSticker}'s \code{\link[hexSticker:sticker]{hexSticker::sticker()}},
\CRANpkg{usethis}'s \code{\link[usethis:use_logo]{usethis::use_logo()}} and \CRANpkg{pkgdown}'s
\code{\link[pkgdown:build_favicons]{pkgdown::build_favicons()}}.

Package asset helpers:
\code{\link[=rog_badge_ropengov]{rog_badge_ropengov()}},
\code{\link[=rog_load_font]{rog_load_font()}}
}
\concept{assets}
