% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rog_actions_pkgdown_branch.R
\name{rog_actions_pkgdown_branch}
\alias{rog_actions_pkgdown_branch}
\title{Creates a GitHub action that deploys on \code{gh-pages} branch}
\usage{
rog_actions_pkgdown_branch(pkg = ".", overwrite = TRUE)
}
\arguments{
\item{pkg}{Path to package.}

\item{overwrite}{Overwrite the action if it was already present.}
}
\description{
The GitHub action created would build a \code{pkgdown} of
your package in the \code{gh-pages} branch of your repo.
}
\seealso{
\code{\link[=rog_build]{rog_build()}}
}
