% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rog-add-template-pkgdown.R
\encoding{UTF-8}
\name{rog_add_template_pkgdown}
\alias{rog_add_template_pkgdown}
\title{Configure a pkgdown site with rogtemplate}
\usage{
rog_add_template_pkgdown()
}
\value{
The function is called for its side effects and returns
\code{invisible(NULL)}.
}
\description{
Add the rOpenGov template configuration to your \verb{_pkgdown.yml}. If no
configuration file exists, create one at \verb{pkgdown/_pkgdown.yml}. Also add the
corresponding paths to \code{.Rbuildignore}.
}
\details{
The function also adds the package website URL
(\verb{https://ropengov.github.io/pkgname/}) and repository URL
(\verb{https://github.com/rOpenGov/pkgname}) to \verb{_pkgdown.yml} and DESCRIPTION if
they are not already present.
}
\examples{
\dontrun{
rog_add_template_pkgdown()
}

}
\seealso{
The \CRANpkg{pkgdown} vignettes
\code{vignette("pkgdown", package = "pkgdown")} and
\code{vignette("linking", package = "pkgdown")}. See also
\CRANpkg{usethis}'s \code{\link[usethis:use_tidy_description]{usethis::use_tidy_description()}}.

Site setup helpers:
\code{\link[=rog_actions_pkgdown_branch]{rog_actions_pkgdown_branch()}},
\code{\link[=rog_build]{rog_build()}}
}
\concept{site}
