% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rog_add_template_pkgdown.R
\encoding{UTF-8}
\name{rog_add_template_pkgdown}
\alias{rog_add_template_pkgdown}
\title{Configure rogtemplate for your pkgdown site}
\usage{
rog_add_template_pkgdown()
}
\value{
The function is called for its side effects and returns
\code{invisible(NULL)}.
}
\description{
Add the rOpenGov template configuration to your \verb{_pkgdown.yml}. If no
configuration file exists, create one at \verb{pkgdown/_pkgdown.yml}. Also add the
corresponding paths to \code{.Rbuildignore}.
}
\details{
The function also adds these package URLs to \verb{_pkgdown.yml} and DESCRIPTION
if they are not already present:
\itemize{
\item https://ropengov.github.io/pkgname/
\item https://github.com/rOpenGov/pkgname
}
}
\examples{
\dontrun{
rog_add_template_pkgdown()
}

}
\seealso{
\CRANpkg{pkgdown} vignettes:
\code{vignette("pkgdown", package = "pkgdown")} and
\code{vignette("linking", package = "pkgdown")},
\CRANpkg{usethis}'s \code{\link[usethis:use_tidy_description]{usethis::use_tidy_description()}}.

site setup helpers:
\code{\link[=rog_actions_pkgdown_branch]{rog_actions_pkgdown_branch()}},
\code{\link[=rog_build]{rog_build()}}
}
\concept{site}
