% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rog_badge_ropengov.R
\encoding{UTF-8}
\name{rog_badge_ropengov}
\alias{rog_badge_ropengov}
\title{Add an rOpenGov badge to your README}
\usage{
rog_badge_ropengov(install = TRUE)
}
\arguments{
\item{install}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, the badge is installed on your
\code{README.md} or \code{README.Rmd}. If \code{FALSE}, a message with the R Markdown code
is displayed.}
}
\value{
The function is called for its side effects and returns \code{NULL}
invisibly.
}
\description{
Add an rOpenGov badge pointing to \url{https://ropengov.org/} to your README:
}
\details{
\ifelse{html}{\href{https://ropengov.org/}{\figure{ropengov-badge.svg}{
options: alt='rOpenGov'}}}{**rOpenGov**}
}
\examples{
rog_badge_ropengov(install = FALSE)
}
\seealso{
\CRANpkg{usethis}'s \code{\link[usethis:use_badge]{usethis::use_badge()}}.

package asset helpers:
\code{\link[=rog_load_font]{rog_load_font()}},
\code{\link[=rog_logo]{rog_logo()}}
}
\concept{assets}
