% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rog_badge_ropengov.R
\name{rog_badge_ropengov}
\alias{rog_badge_ropengov}
\title{Add a rOpenGov badge to your \code{README} file}
\usage{
rog_badge_ropengov(install = TRUE)
}
\arguments{
\item{install}{TRUE/FALSE. On TRUE the badge would be installed on your
README.md/.Rmd. On FALSE a message with the Rmarkdown code would be
displayed.}
}
\value{
A markdown badge or a message with the markdown code to be pasted.
}
\description{
Add an rOpenGov badge pointing to \url{http://ropengov.org/} on your \code{README}
file:

\ifelse{html}{\href{http://ropengov.org/}{\figure{ropengov-badge.svg}{options: alt='rOpenGov'}}}{**rOpenGov**}
}
\examples{

rog_badge_ropengov(install = FALSE)
}
\seealso{
\code{\link[usethis:badges]{usethis::use_badge()}}
}
\concept{extras}
