% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rog-build.R
\encoding{UTF-8}
\name{rog_build}
\alias{rog_build}
\title{Build your pkgdown site locally}
\usage{
rog_build(pkg = ".", ...)
}
\arguments{
\item{pkg}{Path to the package.}

\item{...}{
  Arguments passed on to \code{\link{rog_logo}}, \code{\link[pkgdown:build_site]{pkgdown::build_site}}
  \describe{
    \item{\code{overwrite}}{Should the current logo be overwritten? When \code{TRUE}, it
runs \CRANpkg{usethis}'s \code{\link[usethis:use_logo]{usethis::use_logo()}}.}
    \item{\code{favicons}}{Should favicons be created with \CRANpkg{pkgdown}'s
\code{\link[pkgdown:build_favicons]{pkgdown::build_favicons()}}?}
    \item{\code{p_x}}{x position for package name}
    \item{\code{p_y}}{y position for package name}
    \item{\code{p_size}}{font size for package name}
    \item{\code{examples}}{Run examples?}
    \item{\code{preview}}{If \code{TRUE}, or \code{is.na(preview) && interactive()}, will preview
freshly generated section in browser.}
    \item{\code{new_process}}{If \code{TRUE}, will run \code{build_site()} in a separate process.
This enhances reproducibility by ensuring nothing that you have loaded
in the current process affects the build process.}
    \item{\code{install}}{If \code{TRUE}, will install the package in a temporary library
so it is available for vignettes.}
  }}
}
\value{
The function is called for its side effects and returns
\code{invisible(NULL)}.
}
\description{
A wrapper around \CRANpkg{pkgdown}'s \code{\link[pkgdown:build_site]{pkgdown::build_site()}}.
}
\details{
This function also calls \code{\link[=rog_add_template_pkgdown]{rog_add_template_pkgdown()}} and \code{\link[=rog_logo]{rog_logo()}}.
}
\seealso{
\CRANpkg{pkgdown}'s \code{\link[pkgdown:build_site]{pkgdown::build_site()}}.

Site setup helpers:
\code{\link[=rog_actions_pkgdown_branch]{rog_actions_pkgdown_branch()}},
\code{\link[=rog_add_template_pkgdown]{rog_add_template_pkgdown()}}
}
\concept{site}
