% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rog_build.R
\encoding{UTF-8}
\name{rog_build}
\alias{rog_build}
\title{Build your pkgdown site locally}
\usage{
rog_build(pkg = ".", ...)
}
\arguments{
\item{pkg}{Path to package.}

\item{...}{Arguments passed to \code{\link[=rog_logo]{rog_logo()}} and \CRANpkg{pkgdown}'s
\code{\link[pkgdown:build_site]{pkgdown::build_site()}}. Common arguments for \code{\link[=rog_logo]{rog_logo()}} include
\code{overwrite}, \code{favicons}, \code{p_x}, \code{p_y} and \code{p_size}.}
}
\value{
The function is called for its side effects and returns
\code{invisible(NULL)}.
}
\description{
A wrapper around \CRANpkg{pkgdown}'s \code{\link[pkgdown:build_site]{pkgdown::build_site()}}.
}
\details{
This function also calls \code{\link[=rog_add_template_pkgdown]{rog_add_template_pkgdown()}} and \code{\link[=rog_logo]{rog_logo()}}.
}
\seealso{
\CRANpkg{pkgdown}'s \code{\link[pkgdown:build_site]{pkgdown::build_site()}} and
\code{\link[=rog_add_template_pkgdown]{rog_add_template_pkgdown()}}.

site setup helpers:
\code{\link[=rog_actions_pkgdown_branch]{rog_actions_pkgdown_branch()}},
\code{\link[=rog_add_template_pkgdown]{rog_add_template_pkgdown()}}
}
\concept{site}
