% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rog_build.R
\name{rog_build}
\alias{rog_build}
\title{Build your pkgdown website locally}
\usage{
rog_build(pkg = ".", ...)
}
\arguments{
\item{pkg}{Path to package.}

\item{...}{
  Arguments passed on to \code{\link[=rog_logo]{rog_logo}}, \code{\link[pkgdown:build_site]{pkgdown::build_site}}
  \describe{
    \item{\code{overwrite}}{Should the current logo be overwritten? When \code{TRUE} it
would run \code{\link[usethis:use_logo]{usethis::use_logo()}}.}
    \item{\code{favicons}}{Logical, do you want to create favicons with
\code{\link[pkgdown:build_favicons]{pkgdown::build_favicons()}}?}
    \item{\code{p_x}}{x position for package name}
    \item{\code{p_y}}{y position for package name}
    \item{\code{p_size}}{font size for package name}
    \item{\code{examples}}{Run examples?}
    \item{\code{run_dont_run}}{Run examples that are surrounded in \\dontrun?}
    \item{\code{seed}}{Seed used to initialize random number generation in order to
make article output reproducible. An integer scalar or \code{NULL} for no seed.}
    \item{\code{lazy}}{If \code{TRUE}, will only rebuild articles and reference pages
if the source is newer than the destination.}
    \item{\code{preview}}{If \code{TRUE}, or \code{is.na(preview) && interactive()}, will preview
freshly generated section in browser.}
    \item{\code{devel}}{Use development or deployment process?

If \code{TRUE}, uses lighter-weight process suitable for rapid
iteration; it will run examples and vignettes in the current process,
and will load code with \code{pkgload::load_all()}.

If \code{FALSE}, will first install the package to a temporary library,
and will run all examples and vignettes in a new process.

\code{build_site()} defaults to \code{devel = FALSE} so that you get high fidelity
outputs when you building the complete site; \code{build_reference()},
\code{build_home()} and friends default to \code{devel = TRUE} so that you can
rapidly iterate during development.}
    \item{\code{new_process}}{If \code{TRUE}, will run \code{build_site()} in a separate process.
This enhances reproducibility by ensuring nothing that you have loaded
in the current process affects the build process.}
    \item{\code{install}}{If \code{TRUE}, will install the package in a temporary library
so it is available for vignettes.}
    \item{\code{document}}{\strong{Deprecated} Use \code{devel} instead.}
  }}
}
\description{
A wrapper of \code{\link[pkgdown:build_site]{pkgdown::build_site()}}
}
\details{
This function would call also \code{\link[=rog_add_template_pkgdown]{rog_add_template_pkgdown()}} and
\code{\link[=rog_logo]{rog_logo()}}.
}
\seealso{
\code{\link[pkgdown:build_site]{pkgdown::build_site()}}, \code{\link[=rog_add_template_pkgdown]{rog_add_template_pkgdown()}}.
}
