% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rog_logo.R
\encoding{UTF-8}
\name{rog_load_font}
\alias{rog_load_font}
\title{Load rogtemplate fonts}
\usage{
rog_load_font()
}
\value{
The font family name, \code{"B612 Mono"}.
}
\description{
Load the current rOpenGov font,
\href{https://fonts.google.com/specimen/B612+Mono}{B612 Mono}.
}
\examples{
rog_load_font()
}
\seealso{
package asset helpers:
\code{\link[=rog_badge_ropengov]{rog_badge_ropengov()}},
\code{\link[=rog_logo]{rog_logo()}}
}
\concept{assets}
