% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rog_logo.R
\name{rog_load_font}
\alias{rog_load_font}
\title{Load rogtemplate fonts}
\usage{
rog_load_font()
}
\description{
Load the current font in use for rOpenGov,
\href{https://fonts.google.com/specimen/B612+Mono}{B612 Mono}
}
\examples{
rog_load_font()
}
\seealso{
Other extras: 
\code{\link{rog_add_template_pkgdown}()},
\code{\link{rog_badge_ropengov}()},
\code{\link{rog_logo}()}
}
\concept{extras}
