% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rog_logo.R
\name{rog_logo}
\alias{rog_logo}
\title{Creates a logo for your rOpenGov package}
\usage{
rog_logo(pkgname, filename = "man/figures/logo.png", favicons = TRUE)
}
\arguments{
\item{pkgname}{Name of the package. If not present it would be autodetected
by the function.}

\item{filename}{filename to save sticker}

\item{favicons}{Logical, do you want to create favicons with
\code{\link[pkgdown:build_favicons]{pkgdown::build_favicons()}}?}
}
\description{
Creates a logo for your rOpenGov package
}
\examples{

tmp <- tempfile(fileext = ".png")
rog_logo("test a package", tmp, favicons = FALSE)

# Display
logo <- magick::image_read(tmp)

print(logo)
}
\seealso{
\code{\link[hexSticker:sticker]{hexSticker::sticker()}}, \code{\link[usethis:use_logo]{usethis::use_logo()}},
\code{\link[pkgdown:build_favicons]{pkgdown::build_favicons()}}.
}
\concept{extras

Creates a logo automatically with \code{\link[hexSticker:sticker]{hexSticker::sticker()}}.
Optionally, create also favicons with \code{\link[pkgdown:build_favicons]{pkgdown::build_favicons()}}.}
