% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rog_pals.R
\encoding{UTF-8}
\name{rog_pals}
\alias{rog_pals}
\alias{rog_orange_pal}
\alias{rog_dark_pal}
\alias{rog_gradient_pal}
\alias{rog_qualitative_pal}
\title{rOpenGov color palettes}
\usage{
rog_orange_pal(n = 4, alpha = 0.9, rev = FALSE)

rog_dark_pal(n = 4, alpha = 0.9, rev = FALSE)

rog_gradient_pal(n = 4, alpha = 0.9, rev = FALSE)

rog_qualitative_pal(n = 6, alpha = 0.9, rev = FALSE)
}
\arguments{
\item{n}{Number of colors to generate.}

\item{alpha}{Alpha transparency for the generated colors. Use \code{1} for fully
opaque colors.}

\item{rev}{Should the palette order be reversed?}
}
\value{
Color palettes.
}
\description{
Color palettes based on the rOpenGov color scheme.
}
\examples{
\dontshow{if (requireNamespace("scales", quietly = TRUE)) withAutoprint(\{ # examplesIf}
scales::show_col(rog_orange_pal(9))

scales::show_col(rog_dark_pal(9))

scales::show_col(rog_gradient_pal(9))

scales::show_col(rog_qualitative_pal(9))
\dontshow{\}) # examplesIf}
}
\seealso{
\pkg{grDevices}' \code{\link[grDevices:hcl.colors]{grDevices::hcl.colors()}}.
}
\concept{palettes}
