% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogtemplate-package.R
\docType{package}
\encoding{UTF-8}
\name{rogtemplate-package}
\alias{rogtemplate}
\alias{rogtemplate-package}
\title{rogtemplate: 'pkgdown' Template for 'rOpenGov' Package Sites}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A private pkgdown template for core rOpenGov packages. It provides site setup helpers, package asset helpers and rOpenGov color palettes. Please do not use it for your own code.
}
\seealso{
\code{\link[=rog_add_template_pkgdown]{rog_add_template_pkgdown()}} to configure a site,
\code{\link[=rog_actions_pkgdown_branch]{rog_actions_pkgdown_branch()}} to configure deployment and \code{\link[=rog_build]{rog_build()}}
to build the site locally.
}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Steph Buongiorno \email{steph.buon@gmail.com} (\href{https://orcid.org/0000-0002-6965-0787}{ORCID}) [contributor]
}

}
\keyword{internal}
