% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogtemplate-package.R
\docType{package}
\name{rogtemplate-package}
\alias{rogtemplate}
\alias{rogtemplate-package}
\title{rogtemplate: 'pkgdown' Template for 'rOpenGov' Packages}
\description{
This is a private template for use by core rOpenGov packages.
    Please don’t use for your own code.
}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

}
\keyword{internal}
