% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogtemplate-package.R
\docType{package}
\name{rogtemplate-package}
\alias{rogtemplate}
\alias{rogtemplate-package}
\title{rogtemplate: 'pkgdown' Template for 'rOpenGov' Packages}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This is a private template for use by core rOpenGov packages.
    Please don’t use for your own code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropengov.github.io/rogtemplate/}
  \item \url{https://github.com/ropengov/rogtemplate}
  \item Report bugs at \url{https://github.com/ropengov/rogtemplate/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

}
\keyword{internal}
